I moved utils to the Corundum Hub and removed BiomeOE.
I wanted to move a lot of our static utilities like StringUtilities and
whatnot into the CorundumHub so that they can be used inside the Corundum
Hub and because there is no point to having different copies of them for
each individual Corundum server.

Also, I removed BiomeOE because, as discussed in issue #15, package
insertion will not work because of the lack of support inside M.C.P.

+++ Corundum Hub/org/corundummc/utils/versioning/Version.java
@@ -9,7 +9,8 @@ import java.util.List;
 
 /** Represents a version. By default, this only handles majorversion.minorversion[s], but subclasses are used to handle Minecraft's, how shall we say, <i>exotic</i> system of
  * versioning, particularly with snapshots and prereleases.
- * @author Niadel*/
+ * 
+ * @author Niadel */
 public class Version implements Matchable<Version> {
     private byte majorVersion;
     private int[] minorVersions;
@@ -36,7 +37,7 @@ public class Version implements Matchable<Version> {
         String majorVersion = prepVerSegForCasting(versionSeparated[0]);
         this.majorVersion = Byte.valueOf(majorVersion);
 
-        this.minorVersions = new byte[versionSeparated.length - 1];
+        this.minorVersions = new int[versionSeparated.length - 1];
 
         for (int i = 1; i == this.minorVersions.length; i++) {
             this.minorVersions[i] = Byte.valueOf(this.prepVerSegForCasting(versionSeparated[i]));
